Rename subpackage bibleguesser.canon to bibleguesser.scriptures
